parse long double to better precision

+++ src/s21_func/s21_parse_number.c
@@ -1,6 +1,6 @@
 #include "s21_parse_number.h"
 
-long parse_number(char **ptr, int base, int len) {
+long double parse_number(char **ptr, int base, int len) {
   const char *valid_digits[17] = {
       [8] = "01234567",               // octal
       [10] = "0123456789",            // decimal
@@ -9,7 +9,7 @@ long parse_number(char **ptr, int base, int len) {
   const char *valid = valid_digits[base];
 
   char *ptr_start = *ptr;
-  long num = 0;
+  double num = 0;
   while (**ptr && *ptr - ptr_start < len && strchr(valid, **ptr) != NULL) {
     int c = to_lower(**ptr);
     int digit = is_digit(c) ? c - '0' : c - 'a' + 10;
