+++ src/s21_func/s21_parse_number.h
@@ -4,6 +4,6 @@
 #include "s21_char_utils.h"
 #include "s21_string.h"
 
-long parse_number(char **ptr, int base, int len);
+long double parse_number(char **ptr, int base, int len);
 
 #endif // S21_STRTOL_H
